requirements.txt
streamlit
streamlit-lottie
openai
google-generativeai
requests
PyPDF2
azure-ai-formrecognizer
azure-core
python-dotenv
pandas
folium
streamlit_folium
matplotlib
